Split out ChangesListItem from ChangeScreen and tidy react imports
