+++ app/App.tsx
@@ -1,7 +1,7 @@
 // Import once in application. Global singleton for inversify
 import "reflect-metadata";
 
-import React from "react";
+import React, { Component, ReactNode } from "react";
 import { NavigationContainer } from "react-navigation";
 import { StatusBar, View } from "react-native";
 import Constants from 'expo-constants';
@@ -22,8 +22,8 @@ global.serviceLocator = ConfigureDependencies(appSettingsJson as Appsettings);
 
 const AppContainer: NavigationContainer = createMyAppContainer();
 
-export default class App extends React.Component {
-  public render(): React.ReactNode {
+export default class App extends Component {
+  public render(): ReactNode {
     return (
       <View style={{ flex: 1, paddingTop: Constants.statusBarHeight }}>
         <StatusBar/>
